update nextButton click path change
